pyproject.toml
[tool.poetry]
name = "eddy"
version = "0.1.0"
description = "Fast online clustering for nonstationary data."
authors = ["Michael Hood <7614815+MichaelAHood@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
numpy = "^1.26.0"
faiss-cpu = "^1.7.4"
scipy = "^1.11.3"
networkx = "^3.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pytest = "^7.4.2"
ipython = "^8.15.0"
matplotlib = "^3.8.0"
ruff = "^0.0.291"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
